requirements.txt
numpy
pandas
sklearn
pillow
keras==2.3.1
tensorflow
matplotlib
